requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
boto3==1.35.10
botocore==1.35.10
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Django==4.2.16
django-bootstrap-v5==1.0.11
django-cors-headers==4.4.0
django-storages==1.14.4
idna==3.8
jmespath==1.0.1
load-dotenv==0.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
urllib3==2.2.2
webvtt-py==0.5.1
webvtt-to-json==0.2
